update gradle and maven integration tests
