+++ source-formatter.properties
@@ -215,4 +215,5 @@ source.formatter.excludes=\
     tools/plugins/com.liferay.ide.portlet.core/src/com/liferay/ide/portlet/core/lfportlet/model/LiferayPortlet.java,\
     tools/plugins/com.liferay.ide.portlet.core/src/com/liferay/ide/portlet/core/model/Portlet.java,\
     integration-tests/tools/com.liferay.ide.maven.core.tests/src/com/liferay/ide/maven/core/tests/MavenModuleProjectTests.java,\
+    integration-tests/tools/com.liferay.ide.maven.core.tests/src/com/liferay/ide/maven/core/tests/NewModuleMavenTests.java,\
     functional-tests/tools/workspace/workspace.wizard/workspace.wizard.create.tomcat.70/src/com/liferay/ide/functional/workspace/tests/NewLiferayWorkspaceWizardGradleTests.java
