+++ tools/plugins/com.liferay.ide.core/src/com/liferay/ide/core/workspace/WorkspaceConstants.java
@@ -73,6 +73,8 @@ public class WorkspaceConstants {
 
 	public static final String WARS_DIR_PROPERTY = "liferay.workspace.wars.dir";
 
+	public static final String WORKSPACE_PRODUCT_7_1 = "portal-7.1-ga4";
+
 	public static final Map<String, String[]> liferayTargetPlatformVersions = new HashMap<String, String[]>() {
 		{
 			put("7.3", new String[] {"7.3.2", "7.3.1", "7.3.0"});
